% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreadtxtutil.R
\name{prepareRDSforplotread}
\alias{prepareRDSforplotread}
\title{prepareRDSforplotread}
\usage{
prepareRDSforplotread(rdsdir, os = "macOS")
}
\arguments{
\item{rdsdir}{path to the output rds file}
}
\value{
preperation for plotreading test
}
\description{
Explain later
}
