% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotreadtxtutil.R
\name{saveRDSforplotread}
\alias{saveRDSforplotread}
\title{Load a matrix from a file}
\usage{
saveRDSforplotread(txtdir, rdsdir, os = "macOS")
}
\arguments{
\item{txtdir}{path to the input text file}

\item{rdsdir}{path to the output rds file}
}
\value{
saves rds object for plotreading test
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
